+++ tests/Unit/Messaging/CloudMessageTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Tests\Unit\Messaging;
 
 use Kreait\Firebase\Exception\InvalidArgumentException;
+use Kreait\Firebase\Exception\Messaging\InvalidArgument;
 use Kreait\Firebase\Messaging\CloudMessage;
 use Kreait\Firebase\Messaging\MessageTarget;
 use Kreait\Firebase\Messaging\Notification;
@@ -40,5 +41,37 @@ class CloudMessageTest extends TestCase
     public function testAnEmptyMessageHasNotTarget()
     {
         $this->assertFalse(CloudMessage::new()->hasTarget());
+    }
+
+    /**
+     * @dataProvider multipleTargets
+     */
+    public function testAMessageCanOnlyHaveOneTarget($data)
+    {
+        $this->expectException(InvalidArgument::class);
+        CloudMessage::fromArray($data);
+    }
+
+    public function multipleTargets()
+    {
+        return [
+            'condition and token' => [[
+                MessageTarget::CONDITION => 'something',
+                MessageTarget::TOKEN => 'something else',
+            ]],
+            'condition and topic' => [[
+                MessageTarget::CONDITION => 'something',
+                MessageTarget::TOPIC => 'something else',
+            ]],
+            'token and topic' => [[
+                MessageTarget::TOKEN => 'something',
+                MessageTarget::TOPIC => 'something else',
+            ]],
+            'all of them' => [[
+                MessageTarget::CONDITION => 'something',
+                MessageTarget::TOKEN => 'something else',
+                MessageTarget::TOPIC => 'something even elser',
+            ]],
+        ];
     }
 }